Fix SPJ query Merging
select-project-join queries can be flattened. But we currently encode
GROUP BY's as a wrapper around an SPJ, so after the flattening the
groupby expression doesn't get properly substituted.

Fix: track the substitutions and do a final pass to apply them.
Ugly, but works for now. Ideally find a nicer way to express groupbys
and similar wrappers in a way that respects scoping, but we do not want
the groupby ops into the result tuple either so scoping is icky either
way.

We could add the groupby (and sorting and whatever else) into the spj
record and add pattern synonyms to check if we have a 'real' spj. Sort
of works i guess?